# apps/mcp-hub/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Truline MCP Hub",
    "description": "Personal OS tool server. Exposes homelab, calendar, and finance tools as OpenAPI for Open WebUI.",
    "version": "0.1.1"
  },
  "paths": {
...
    "/api/v1/factory/task_runnable": {
      "post": {
        "tags": [
          "Factory"
        ],
        "summary": "Whether a dev.task can be claimed right now, and why not if it can't",
        "operationId": "get_task_runnable",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRunnableQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Task Runnable"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/factory/probe": {
      "get": {
        "tags": [
          "Factory"
        ],
        "summary": "Probe the console operator surface: read what the release index and factory-state panel claim, re-read the store and the schedule independently, and record the comparison (R26.09/O-7)",
        "operationId": "get_console_surface_probe",
        "parameters": [
          {
            "name": "release_ref",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Release ref(s) to check; omit for every non-terminal release",
              "title": "Release Ref"
            },
            "description": "Release ref(s) to check; omit for every non-terminal release"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Console Surface Probe"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/factory/impact": {
      "post": {
        "tags": [
          "Factory"
        ],
        "summary": "What a change, application or CI affects, over the graph's typed edges (affects/depends_on/consumes) -- with the coverage the answer is computed on, stated beside it (R26.09/O-4, PC-ASR-002/AC-1)",
        "operationId": "get_impact_analysis",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpactQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Impact Analysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/factory/tasks": {
      "post": {
        "tags": [
          "Factory"
        ],
        "summary": "File a dev.task bead from a JSON spec \u2014 the same shape file_task.py's CLI accepts",
        "operationId": "file_dev_task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Spec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response File Dev Task"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/factory/findings": {
      "post": {
        "tags": [
          "Factory"
        ],
        "summary": "File a dev.finding bead from a JSON spec \u2014 the same shape file_finding.py's CLI accepts",
        "operationId": "file_dev_finding",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Spec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response File Dev Finding"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/factory/note": {
      "post": {
        "tags": [
          "Factory"
        ],
        "summary": "File a dev.note bead against a dev.task -- comments, answers, status, review",
        "operationId": "file_dev_note",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response File Dev Note"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/factory/prs/{number}/merge": {
      "post": {
        "tags": [
          "Factory"
        ],
        "summary": "Start a merge-by-verdict workflow for a PR (R26.09/O-2) -- ships dark until a decision record grants factory.merge; see .factory/design.md",
        "operationId": "start_pr_merge_on_verdict",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Start Pr Merge On Verdict"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/factory/prs/merge/{workflow_id}": {
      "get": {
        "tags": [
          "Factory"
        ],
        "summary": "Read a merge-by-verdict workflow's recorded disposition",
        "operationId": "get_pr_merge_disposition",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Pr Merge Disposition"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImpactQuery": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "change",
              "application",
              "ci"
            ],
            "title": "Kind",
            "description": "The EA object type to compute the impact of"
          },
          "ref": {
            "type": "string",
            "minLength": 1,
            "title": "Ref",
            "description": "content.ref of the arch.<kind> bead"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "ref"
        ],
        "title": "ImpactQuery"
      },
...
      "NoteCreateRequest": {
        "properties": {
          "parent_id": {
            "type": "string",
            "title": "Parent Id",
            "description": "dev.task bead id this note is filed against"
          },
          "kind": {
            "type": "string",
            "enum": [
              "comment",
              "question",
              "answer",
              "status",
              "attachment",
              "review"
            ],
            "title": "Kind"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "blocking": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocking"
          },
          "answers_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answers Ref"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "approve",
                  "request-changes"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "releases_work": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Releases Work"
          }
        },
        "type": "object",
        "required": [
          "parent_id",
          "kind",
          "body"
        ],
        "title": "NoteCreateRequest",
        "description": "The closed dev.note shape -- mirrors\napps/substrate/src/schemas.py::DevNoteContent's own implied shape, so\nthis route refuses exactly what the store would refuse (this bead's\nAC-3), not a second intake that accepts more and just moves the 422 one\nlayer out."
      },
...
      "TaskRunnableQuery": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "dev.task bead id"
          }
        },
        "type": "object",
        "required": [
          "task_id"
        ],
        "title": "TaskRunnableQuery"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}